add changes in views.py file
